Streamline heroku deployment process
Now deploying straight from the master branch. Added functionality for
a custom environmental variable (HEROKU_EXAMPLE). If set to true,
the application will check whether a test user account exists and if it
doesn't - it will create one. The HEROKU_EXAMPLE is also exposed as a
app.locals variable which, if set to true, adds Google Analytics script
to the default Handlebars layout and displays test user account
information on the login page.