Fix failing tests, change MONGOURI to MONGO_URI, fix local testing
One of the frontend - task test was failing, causing few others to fail
as a consequence. It turned out that it takes some time until
a particular selector gets updated, especially when connected to a
remote database.

MONGOURI environmental variable got updated to MONGO_URI.

Fixed backend tests when using a local mongo connection. It could not
find the necessary config file with the local mongo URI details.