Rework logging mechanism
Logger now logs both info and error-level messages to the console
and messages.log file on production. On development everything goes
to the console.

Morgan now logs to the console and access.log file on production.
logrotate-stream makes sure that there are only up to 6 log files, 100k
each. On development it only logs to the console.