Rework tests, change babel dependency to babel-core
Tests now can be run with `npm test`. It uses gulp to start up a local
server instance and then it runs mocha and mocha-casperjs to do
the testing.

Added casper.options.remoteScripts with the polyfill.io script.

Few other small fixes:
- removed phantomjs local dependency
(now it is required to have it installed globally, as per README),
- removed gulp-mocha and forever dependency
- moved chalk from devDependencies to dependencies, as it is required
on production