Rework server initialization, few cosmetic fixes
app.js now exports a Server class which supports events.
bootstrap.js instantiates a server object and uses `connect`
function to start the server. When running in development (with gulp),
server events are used to set up browser-sync to reload the server
and/or the browser when necessary. gulp-nodemon dependency was dropped.