Fix server instantiation
Implementation introduced in fe6fb8cfb2d26824aeabaf589a98f50eb7f8081c
didn't allow for restarting the server completely when a file changed.
While this wasn't an issue on production or testing environments,
since we don't hot-reload the server and refresh the browser on demand
there, any change introduced to server files wasn't reflected on the
particular server instance that the browser was connected to.

This implementation utilizes child_process.fork which allows to
spawn a separate node process and kill it when necessary.